+++ src/functionalTest/java/io/knotx/stack/functional/GatewayAPIIntegrationTest.java
@@ -13,17 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.stack.it;
+package io.knotx.stack.functional;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.get;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import io.knotx.junit5.KnotxApplyConfiguration;
 import io.knotx.junit5.KnotxExtension;
 import io.knotx.junit5.RandomPort;
+import io.knotx.junit5.util.FileReader;
 import io.knotx.junit5.wiremock.ClasspathResourcesMockServer;
+import io.knotx.stack.KnotxServerTester;
+import io.vertx.core.json.JsonObject;
 import io.vertx.junit5.VertxTestContext;
 import io.vertx.reactivex.core.Vertx;
 import org.junit.jupiter.api.DisplayName;
@@ -34,34 +35,24 @@ import org.junit.jupiter.api.extension.ExtendWith;
 
 @ExtendWith(KnotxExtension.class)
 @TestInstance(Lifecycle.PER_CLASS)
-class LongRunningHttpServiceWithCircuitBreakerIntegrationTest {
+class GatewayAPIIntegrationTest {
 
   @ClasspathResourcesMockServer
   private WireMockServer mockService;
 
-  @ClasspathResourcesMockServer
-  private WireMockServer mockRepository;
-
   @Test
-  @DisplayName("Expect page containing data from services and fallback data for broken service.")
+  @DisplayName("Expect response containing data from HTTP service")
   @KnotxApplyConfiguration({"conf/application.conf",
-      "scenarios/long-running-http-service-with-circuit-breaker/mocks.conf",
-      "scenarios/long-running-http-service-with-circuit-breaker/tasks.conf"})
-  void taskWithCircuitBreaker(VertxTestContext context, Vertx vertx,
-      @RandomPort Integer delayedServicePort,
-      @RandomPort Integer globalServerPort) {
-    // given
-    WireMockServer wireMockServer = new WireMockServer(delayedServicePort);
-    wireMockServer.stubFor(get(urlEqualTo("/service/mock/delayed")).willReturn(
-        aResponse()
-            .withStatus(200)
-            .withFixedDelay(2000)));
-
-    // when
+      "scenarios/gateway-api-with-fragments/mocks.conf",
+      "scenarios/gateway-api-with-fragments/tasks.conf"})
+  void requestPage(VertxTestContext context, Vertx vertx, @RandomPort Integer globalServerPort) {
     KnotxServerTester serverTester = KnotxServerTester.defaultInstance(globalServerPort);
-    serverTester
-        .testGetRequest(context, vertx, "/content/fullPage.html",
-            "scenarios/long-running-http-service-with-circuit-breaker/result/fullPage.html");
+    serverTester.testGet(context, vertx, "/api/author-info",
+        resp -> {
+          String expectedResponse = FileReader
+              .readTextSafe("scenarios/gateway-api-with-fragments/result/author-info.json");
+          assertEquals(new JsonObject(expectedResponse), resp.body().toJsonObject());
+        });
   }
 
 }