+++ src/functionalTest/java/io/knotx/stack/functional/LongRunningHttpServiceWithCircuitBreakerIntegrationTest.java
@@ -13,21 +13,30 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.stack.it;
+package io.knotx.stack.functional;
+
+import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.get;
+import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import io.knotx.junit5.KnotxApplyConfiguration;
 import io.knotx.junit5.KnotxExtension;
 import io.knotx.junit5.RandomPort;
 import io.knotx.junit5.wiremock.ClasspathResourcesMockServer;
+import io.knotx.stack.KnotxServerTester;
 import io.vertx.junit5.VertxTestContext;
 import io.vertx.reactivex.core.Vertx;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.junit.jupiter.api.extension.ExtendWith;
 
+
 @ExtendWith(KnotxExtension.class)
-class HttpServiceIntegrationTest {
+@TestInstance(Lifecycle.PER_CLASS)
+class LongRunningHttpServiceWithCircuitBreakerIntegrationTest {
 
   @ClasspathResourcesMockServer
   private WireMockServer mockService;
@@ -36,14 +45,25 @@ class HttpServiceIntegrationTest {
   private WireMockServer mockRepository;
 
   @Test
-  @DisplayName("Expect page containing data from HTTP services.")
+  @DisplayName("Expect page containing data from services and fallback data for broken service.")
   @KnotxApplyConfiguration({"conf/application.conf",
-      "scenarios/http-service/mocks.conf",
-      "scenarios/http-service/tasks.conf"})
-  void requestPage(VertxTestContext context, Vertx vertx,
+      "scenarios/long-running-http-service-with-circuit-breaker/mocks.conf",
+      "scenarios/long-running-http-service-with-circuit-breaker/tasks.conf"})
+  void taskWithCircuitBreaker(VertxTestContext context, Vertx vertx,
+      @RandomPort Integer delayedServicePort,
       @RandomPort Integer globalServerPort) {
+    // given
+    WireMockServer wireMockServer = new WireMockServer(delayedServicePort);
+    wireMockServer.stubFor(get(urlEqualTo("/service/mock/delayed")).willReturn(
+        aResponse()
+            .withStatus(200)
+            .withFixedDelay(2000)));
+
+    // when
     KnotxServerTester serverTester = KnotxServerTester.defaultInstance(globalServerPort);
-    serverTester.testGetRequest(context, vertx, "/content/fullPage.html", "results/fullPage.html");
+    serverTester
+        .testGetRequest(context, vertx, "/content/fullPage.html",
+            "scenarios/long-running-http-service-with-circuit-breaker/result/fullPage.html");
   }
 
 }