+++ src/functionalTest/java/io/knotx/stack/functional/ResponseHeadersTest.java
@@ -13,44 +13,44 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.knotx.stack.it;
+package io.knotx.stack.functional;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
 import static com.github.tomakehurst.wiremock.client.WireMock.get;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlMatching;
 import static io.knotx.junit5.wiremock.KnotxWiremockExtension.stubForServer;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import io.knotx.junit5.KnotxApplyConfiguration;
 import io.knotx.junit5.KnotxExtension;
 import io.knotx.junit5.RandomPort;
 import io.knotx.junit5.wiremock.ClasspathResourcesMockServer;
+import io.knotx.stack.KnotxServerTester;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.vertx.junit5.VertxTestContext;
+import io.vertx.reactivex.core.MultiMap;
 import io.vertx.reactivex.core.Vertx;
-import org.junit.jupiter.api.BeforeAll;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.TestInstance;
-import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.junit.jupiter.api.extension.ExtendWith;
 
-// TODO remove this test after moving scenario to knotx-http-server repository
 @ExtendWith(KnotxExtension.class)
-@TestInstance(Lifecycle.PER_CLASS)
-class TemplatingIntegrationTest {
+class ResponseHeadersTest {
+
+  private MultiMap expectedHeaders = MultiMap.caseInsensitiveMultiMap();
 
   @ClasspathResourcesMockServer
   private WireMockServer mockService;
 
   @ClasspathResourcesMockServer
-  private WireMockServer mockBrokenService;
-
-  @ClasspathResourcesMockServer
   private WireMockServer mockRepository;
 
-  @BeforeAll
-  void initMocks() {
+  @BeforeEach
+  void before() {
     stubForServer(mockService,
         get(urlMatching("/service/mock/.*"))
             .willReturn(
@@ -60,35 +60,38 @@ class TemplatingIntegrationTest {
                     .withHeader("X-Server", "Knot.x")
             ));
 
-    stubForServer(mockBrokenService,
-        get(urlMatching("/service/broken/.*"))
-            .willReturn(
-                aResponse()
-                    .withStatus(HttpResponseStatus.INTERNAL_SERVER_ERROR.code())
-            ));
-
     stubForServer(mockRepository,
         get(urlMatching("/content/.*"))
             .willReturn(
                 aResponse()
                     .withHeader("Cache-control", "no-cache, no-store, must-revalidate")
                     .withHeader("Content-Type", "text/html; charset=UTF-8")
-                    .withHeader("X-Server", "Knot.x")
+                    .withHeader("X-Server", "Knot.x-Custom-Header")
             ));
+
+    expectedHeaders.clear();
+    expectedHeaders.add("Content-Type", "text/html; charset=UTF-8");
+    expectedHeaders.add("X-Server", "Knot.x-Custom-Header");
   }
 
-  // TODO move to https://github.com/Knotx/knotx-server-http
+  @Disabled("This test actually should be moved to http-server. It only checks if the server properly computes content-length and content-type headers.")
   @Test
-  @DisplayName("Expect page containing data from services when parameters specified.")
-  @KnotxApplyConfiguration({"conf/application.conf",
-      "scenarios/http-service/mocks.conf",
-      "scenarios/http-service/tasks.conf"})
-  void requestPageWithRequestParameters(VertxTestContext context, Vertx vertx,
-      @RandomPort Integer globalServerPort) {
+  @DisplayName("Expect allowed headers in Server response.")
+  @KnotxApplyConfiguration({"conf/application.conf", "scenarios/response-headers/mocks.conf"})
+  void whenRequestingRemoteRepository_expectOnlyAllowedResponseHeaders(VertxTestContext context,
+      Vertx vertx, @RandomPort Integer globalServerPort) {
     KnotxServerTester serverTester = KnotxServerTester.defaultInstance(globalServerPort);
-    serverTester.testGetRequest(context, vertx,
-        "/content/fullPage.html?parameter%20with%20space=value&q=knotx",
-        "results/fullPage.html");
+    serverTester.testGet(context, vertx, "/content/fullPage.html", resp -> {
+      MultiMap headers = resp.headers();
+      expectedHeaders.names().forEach(name -> {
+        assertTrue(headers.contains(name), "Header " + name + " is expected to be present.");
+        assertEquals(expectedHeaders.get(name), headers.get(name),
+            "Wrong value of " + name + " header.");
+      });
+      assertEquals(HttpResponseStatus.OK.code(), resp.statusCode(), "Wrong status code received.");
+      assertTrue(headers.contains("content-length"),
+          "content-length header is expected to be present.");
+    });
   }
 
 }