+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to `knotx-stack` will be documented in this file.
 
 ## Unreleased
 List of changes that are finished but not yet released in any final version.
+- [PR-64](https://github.com/Knotx/knotx-stack/pull/64) - Add Unit tests for default stack configuration files.
 - [PR-59](https://github.com/Knotx/knotx-stack/pull/59) - Vert.x upgrade to 3.8.1.
 
 ## 2.0.0